+++ sdk/core/azure-core/tests/specs_sdk/modeltypes/modeltypes/_utils/model_base.py
@@ -28,7 +28,7 @@ import isodate
 from azure.core.exceptions import DeserializationError
 from azure.core import CaseInsensitiveEnumMeta
 from azure.core.pipeline import PipelineResponse
-from azure.core.serialization import _Null
+from azure.core.serialization import _Null, TypeHandlerRegistry
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -36,6 +36,9 @@ __all__ = ["SdkJSONEncoder", "Model", "rest_field", "rest_discriminator"]
 
 TZ_UTC = timezone.utc
 _T = typing.TypeVar("_T")
+
+
+TYPE_HANDLER_REGISTRY = TypeHandlerRegistry()
 
 
 def _timedelta_as_isostr(td: timedelta) -> str:
@@ -161,6 +164,10 @@ class SdkJSONEncoder(JSONEncoder):
             except AttributeError:
                 # This will be raised when it hits value.total_seconds in the method above
                 pass
+
+            custom_serializer = TYPE_HANDLER_REGISTRY.get_serializer(o)
+            if custom_serializer:
+                return custom_serializer(o)
             return super(SdkJSONEncoder, self).default(o)
 
 
@@ -316,7 +323,10 @@ def get_deserializer(annotation: typing.Any, rf: typing.Optional["_RestField"] =
         return _deserialize_int_as_str
     if rf and rf._format:
         return _DESERIALIZE_MAPPING_WITHFORMAT.get(rf._format)
-    return _DESERIALIZE_MAPPING.get(annotation)  # pyright: ignore
+    func = _DESERIALIZE_MAPPING.get(annotation)  # pyright: ignore
+    if func:
+        return func
+    return TYPE_HANDLER_REGISTRY.get_deserializer(annotation)
 
 
 def _get_type_alias_type(module_name: str, alias_name: str):
@@ -481,6 +491,7 @@ def _is_model(obj: typing.Any) -> bool:
 
 
 def _serialize(o, format: typing.Optional[str] = None):  # pylint: disable=too-many-return-statements
+
     if isinstance(o, list):
         return [_serialize(x, format) for x in o]
     if isinstance(o, dict):
@@ -510,6 +521,12 @@ def _serialize(o, format: typing.Optional[str] = None):  # pylint: disable=too-m
     except AttributeError:
         # This will be raised when it hits value.total_seconds in the method above
         pass
+
+    # Check if there's a custom serializer for the type
+    custom_serializer = TYPE_HANDLER_REGISTRY.get_serializer(o)
+    if custom_serializer:
+        return custom_serializer(o)
+
     return o